코드 리팩토링, 회고 작성
