+++ .github/pull_request_template.md
@@ -4,21 +4,22 @@
 
 - [X] shallowEquals 구현 완료
 - [X] deepEquals 구현 완료
-- [ ] memo 구현 완료
-- [ ] deepMemo 구현 완료
+- [x] memo 구현 완료
+- [x] deepMemo 구현 완료
 - [x] useRef 구현 완료
 - [x] useMemo 구현 완료
-- [ ] useDeepMemo 구현 완료
-- [ ] useCallback 구현 완료
+- [x] useDeepMemo 구현 완료
+- [x] useCallback 구현 완료
 
 ### 심화 과제
 
-- [ ] 기본과제에서 작성한 hook을 이용하여 렌더링 최적화를 진행하였다.
-- [ ] Context 코드를 개선하여 렌더링을 최소화하였다.
+- [x] 기본과제에서 작성한 hook을 이용하여 렌더링 최적화를 진행하였다.
+- [x] Context 코드를 개선하여 렌더링을 최소화하였다.
 
 ## 과제 셀프회고
-
 <!-- 과제에 대한 회고를 작성해주세요 -->
+
+- 이번 과제를 진행하면서 타입스크립트의 기초가 너무 부족하다는 것을 알게되었다. 과제를 진행하는데에 어려움이 있을 정도의 기초가 부실...
 
 ### 기술적 성장
 <!-- 예시
@@ -26,6 +27,13 @@
 - 기존 지식의 재발견/심화
 - 구현 과정에서의 기술적 도전과 해결
 -->
+- shallowEquals와 deepEquals를 구현하면서, JavaScript와 TypeScript의 기본 문법에 대해서 스킬업 된 것 같다.
+  1) null의 타입이 객체(object)이므로 추가로 예외처리가 필요하다
+  2) 객체 순회 시 사용되는 for...in과 for...of의 타입 추론 방식의 차이를 이해했다. for...in은 객체의 속성을 직접 순회하므로 타입스크립트가 자동으로 객체의 키 타입을 추론할 수 있지만, for...of로 Object.keys()를 순회할 경우 반환된 키 배열과 원본 객체의 연결성을 타입스크립트가 보장할 수 없어 명시적인 타입 단언이 필요하다는 것을 알게 되었다.
+
+
+- useRef, useMemo, useCallback 가 어떻게 동작하는 지, 렌더링을 최적화하기 위해 어떻게 사용해야 하는지 알게되어 좋았습니다.
+- Context가 무엇인지, 커스텀 훅에 대해서 배우게 되었습니다. 아직 완벽히 이해하지 못하였지만 어떤 동작원리인지, 렌더링의 라이프사이클에 대해 알게되었습니다.
 
 ### 코드 품질
 <!-- 예시
@@ -34,12 +42,23 @@
 - 코드 설계 관련 고민과 결정
 -->
 
+- 화면에 필요한 컴포넌트와 커스텀훅을 분리하였다. 이 과정에서 관심사를 분리하는 방법을 배울 수 있었습니다.
+
 ### 학습 효과 분석
 <!-- 예시
 - 가장 큰 배움이 있었던 부분
 - 추가 학습이 필요한 영역
 - 실무 적용 가능성
 -->
+
+- useMemo, useCallback 같은 메모이제이션 기법의 존재는 있었지만, 실무에서 사용할 정도의 이해는 없었다. 이번 과제를 계기로 어느 상황에서 사용해야 하는지 알게 되어 좋았습니다
+
+#### 추가 학습이 필요한 영역
+- 여전히 1, 2주차와 마찬가지로 언어의 기본 지식이 부족함을 느꼈습니다. 시간을 내어 JavaScript와 TypeScript 기본문법과 알고리즘 공부를 병행해야 할 것 같습니다.
+  1) 스파르타코딩클럽의 JavaScript, TypeScript 학습
+  2) 백준 / 프로그래머스 코딩테스트 병행 (내장 메서드를 너무 모르는 것 같음...)
+
+- 이후에 코드 리팩토링과 학습내용을 복기하면서 실무에서도 사용할 수 있도록 노력해보겠습니다.
 
 ### 과제 피드백
 <!-- 예시
